Bumping version to 1.0.11.rc.011. Previous Commit message: "add backwards compatibility code for barcode scanner add-on [touch: 10910]
There is documentation that explicitly instructs peopel to use the `ee_edit_checkin` capability for controlling access to the barcode scanner add-on in the admin.  The problem here is that `ee_edit_checkin` capability is a meta capbility and with the work done in the core https://events.codebasehq.com/projects/event-espresso/tickets/10569 ticket, meta capabilities no longer work as expected if the role has that capability and the `current_user_can` check is not sending along an id for checking with.  So although the changes in the conditional are sufficient for against current ee core master, they are not sufficient against the 10569 branch.

Hence, I’ve also implemented temporarily removing the EE_Capabilities meta caps filter before doing the current_user_can check so `ee_edit_checkin` is treated like a “normal” capability for the purpose of this specific check, and then adding the filter back after the check.

This is fugly, but in order to preserve backward compability for users who may have created custom roles with JUST the `ee_edit_checkin` capability and not also adding the `ee_edit_checkins` capability will still see things work as expected.
"